requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
asgiref==3.8.1
assemblyai==0.37.0
asttokens==3.0.0
asyncer==0.0.7
asyncio==3.4.3
attrs==23.2.0
audioread==3.0.1
autobahn==24.4.2
Automat==22.10.0
beartype==0.18.5
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.45.1
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.16.0
chainlit==2.0.5
channels==4.1.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.8
clusterops==0.1.8
colorama==0.4.6
CoLT5-attention==0.11.1
comtypes==1.4.9
configparser==7.0.0
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.10
daphne==4.1.2
dataclasses-json==0.6.7
datasets==3.2.0
decorator==5.2.1
Deprecated==1.2.15
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.4
django-allauth==64.1.0
django-bootstrap4==24.3
django-cleanup==8.0.0
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
elevenlabs==0.2.27
executing==2.2.0
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.6
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
genai==2.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hyper-connections==0.1.9
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
ipython==8.34.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.15
langchain-community==0.0.29
langchain-core==0.3.31
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.1.147
Lazify==0.4.0
lazy_loader==0.4
librosa==0.11.0
litellm==1.59.6
literalai==0.1.103
llama-cloud==0.1.10
llama-index==0.12.13
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.14
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.14
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.20
llvmlite==0.44.0
local-attention==1.11.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
ollama==0.4.5
openai==0.27.10
openai-whisper==20240930
opencv-python==4.10.0.84
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.14
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathos==0.3.3
Pillow==10.1.0
pipwin==0.5.2
platformdirs==4.3.6
playsound==1.2.2
pluggy==1.5.0
pooch==1.8.2
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pypiwin32==223
PyPrind==2.11.3
pyreadline3==3.4.1
pySmartDL==1.3.4
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
python-telegram-bot==21.3
pyttsx3==2.98
pytz==2024.1
pywin32==308
PyYAML==6.0.2
ray==2.41.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.21.0
sentencepiece==0.2.0
service-identity==24.1.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
SpeechRecognition==3.14.0
SQLAlchemy==2.0.37
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
swarm-models==0.3.1
swarms==6.9.7
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
Telethon==1.36.0
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.3.3
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.20.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trio==0.25.1
trio-websocket==0.11.1
twilio==8.10.0
Twisted==24.7.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
uptrace==1.28.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.0
vector-quantize-pytorch==1.17.1
vision-llama==0.0.8
watchfiles==0.20.0
wcwidth==0.2.13
websockets==14.2
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zetascale==2.7.9
zipp==3.21.0
zope.interface==7.0.1
